Added new entities to BTP extractor
Added new entities to BTP extractor
