+++ db/Analytical/contracts/ContractLineItemDocument.cds
@@ -1,7 +1,7 @@
 namespace sap.ariba;
 using { managed } from '@sap/cds/common';
 
-using sap.ariba.type as types from '../types';
+using sap.ariba.type as types from '../../types';
 
 /**
     Name:	                Contract Line Item Document
