+++ db/Analytical/general/UserAbility.cds
@@ -1,6 +1,6 @@
 namespace sap.ariba;
 using { managed } from '@sap/cds/common';
-using sap.ariba.type as types from '../types';
+using sap.ariba.type as types from '../../types';
 
 /**
   Name:        User Ability Fact
