+++ db/Analytical/sourcing/SRProjectInfo.cds
@@ -1,6 +1,6 @@
 namespace sap.ariba;
 using { managed, cuid } from '@sap/cds/common';
-using sap.ariba.type as types from '../types';
+using sap.ariba.type as types from '../../types';
 
 /**
     Name:           Sourcing Request Project Info
