+++ resources/AnalyticalAPIView/sourcing/EXT_Document.json
@@ -1,0 +1,31 @@
+{
+
+    "status": "published",
+    "documentType": "DocumentFact",
+    "selectAttributes": [
+        "InternalId",
+        "Status",
+        "Description",
+        "DocumentType",
+        "Title",
+        "ProjectId",
+        "SourceSystem",
+        "TimeCreated",
+        "TimeUpdated",
+        "CreateDate"
+    ],
+    "filterExpressions": [
+        {
+            "name": "updatedDateTo",
+            "field": "TimeUpdated",
+            "op": "<=",
+            "defaultValue": null
+        },
+        {
+            "name": "updatedDateFrom",
+            "field": "TimeUpdated",
+            "op": ">",
+            "defaultValue": null
+        }
+    ]
+}
